+++ lib/src/shared/services/biometric_service.dart
@@ -1,0 +1,33 @@
+import 'package:local_auth/local_auth.dart';
+
+class BiometricService {
+  Future<bool> authenticate() async {
+    try {
+      var localAuth = LocalAuthentication();
+
+      bool canCheckBiometrics = await localAuth.canCheckBiometrics;
+
+      if (canCheckBiometrics) {
+        List<BiometricType> availableBiometrics =
+            await localAuth.getAvailableBiometrics();
+
+        if (availableBiometrics.isNotEmpty) {
+          bool isAuthenticated = await localAuth.authenticate(
+            localizedReason: 'Authenticate to access the app',
+            options: const AuthenticationOptions(
+                useErrorDialogs: true, stickyAuth: false, biometricOnly: false),
+          );
+
+          return isAuthenticated;
+        } else {
+          return false;
+        }
+      } else {
+        return true;
+      }
+    } catch (e) {
+      print('Error during biometric authentication: $e');
+      return false;
+    }
+  }
+}
